requirements.txt
mkdocs==1.0.4
GitPython==3.1.0
